Update execution time and model appearance in image_classification notebook
The Jupyter notebook for image classification has been updated with new execution times. The modified times relate to parts of the code that generate and display a quantum convolutional neural network model. This model, classified_with_qnn_jax, is used for image classification. The notebook updates do not affect the functionality of the notebook but ensure that example execution times are accurate.